+++ CHANGELOG.md
@@ -9,6 +9,18 @@
 
 ### Removed
 
+## [7.0.0] - 2022-10-13
+### Changed
+- [ext] convert js to mjs
+- 更新 agent 版本
+- 更新 第三方程式庫版本
+
+### Fixed
+- (template) 修正 資源檔路徑
+
+### Removed
+- [rollup config] 移除 __dirname
+
 ## [7.0.0] - 2022-10-07
 ### Added
 - (makeres) 切割子任務,單獨處理指定的子場景 
@@ -16,7 +28,7 @@
 ## [7.0.0] - 2022-10-07
 ### Added
 - (makeres) 新增 參數 resName, 指定處理單一資源
-- template 新增 資源檔參數 vendor.langList 與 vendor.sceneList
+- (template) 新增 資源檔參數 vendor.langList 與 vendor.sceneList
 
 ### Changed
 - 語言代碼 vendor.custom 改為 vendor.langList